Remove all orphan instances
* Remove all orphan instances.

* One of the `unions` definitions used `foldl`; make it use `foldl'`
  instead.

* Remove false caveat about amortized bounds. They definitely *do*
  hold in a persistent setting, and we're very careful to maintain
  that! Someone must have confused lazy-spined binomial queues with
  strict-spined ones.
